food_recomendation/pyproject.toml
[project]
name = "food-recomendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "tensorflow>=2.19.0",
    "tensorflow-hub>=0.16.1",
]
